move editor to components
