+++ frontend/src/types/collection.ts
@@ -1,0 +1,7 @@
+export type Collection = {
+  id: string;
+  oid: string;
+  name: string;
+  description: string | null;
+  crawlIds: string[];
+};
